fix: Pint fully-qualified @mixin; unify portal progress into posting cards
- JobPostingResource: import Enlistment for the @mixin (Pint fully_qualified_strict_types).
- Job Portal: drop the separate "Your applications" panel; show each account's
  progress + reviewer-action timeline inline on the corporation's posting card
  (apply form when not yet applied). Removes the now-redundant
  appliedCorporationIds prop/method and MyApplicationCard.
