+++ resources/js/Pages/Recruitment/Portal/PostingCard.vue
@@ -43,14 +43,60 @@
     </div>
 
     <div class="mt-auto pt-2">
-      <div
-        v-if="applied"
-        class="inline-flex items-center text-sm font-medium text-green-700"
-      >
-        <CheckCircleIcon class="h-5 w-5 mr-1.5" />
-        Applied
-      </div>
+      <!-- Applied: show this account's progress + reviewer-action timeline for this corporation. -->
+      <template v-if="application">
+        <div class="flex items-center justify-between">
+          <span
+            class="inline-flex items-center rounded-full px-2 py-1 text-xs font-medium"
+            :class="statusClass"
+          >
+            {{ statusLabel }}
+          </span>
+          <span
+            v-if="application.status === 'open' && application.total_stages > 0"
+            class="text-xs text-gray-500"
+          >
+            Stage {{ application.current_position + 1 }} of {{ application.total_stages }}: {{ application.current_stage }}
+          </span>
+        </div>
 
+        <div
+          v-if="application.status === 'open' && application.total_stages > 0"
+          class="mt-1 h-1.5 w-full rounded-full bg-gray-100"
+        >
+          <div
+            class="h-1.5 rounded-full bg-indigo-500"
+            :style="{ width: progressPercent + '%' }"
+          />
+        </div>
+
+        <ol
+          v-if="application.timeline.length > 0"
+          class="mt-2 space-y-1 border-t border-gray-100 pt-2"
+        >
+          <li
+            v-for="(entry, index) in application.timeline"
+            :key="index"
+            class="flex items-center gap-2 text-sm"
+          >
+            <CheckCircleIcon
+              v-if="entry.decision === 'accepted'"
+              class="h-4 w-4 shrink-0 text-green-500"
+            />
+            <XCircleIcon
+              v-else
+              class="h-4 w-4 shrink-0 text-red-500"
+            />
+            <span class="text-gray-900">{{ entry.stage_label }}</span>
+            <span class="text-gray-500">
+              {{ entry.decision === 'accepted' ? 'passed' : 'rejected' }} by {{ entry.reviewer }}
+            </span>
+            <span class="ml-auto text-xs text-gray-400">{{ entry.at }}</span>
+          </li>
+        </ol>
+      </template>
+
+      <!-- Not applied: apply form. -->
       <form
         v-else
         class="space-y-2"
@@ -91,7 +137,7 @@
 
 <script setup>
 import { useForm } from "@inertiajs/vue3";
-import { CheckCircleIcon } from "@heroicons/vue/24/solid";
+import { CheckCircleIcon, XCircleIcon } from "@heroicons/vue/24/solid";
 import { computed } from "vue";
 import EveImage from "@/Shared/EveImage.vue";
 
@@ -104,9 +150,10 @@ const props = defineProps({
     required: true,
     type: Array,
   },
-  applied: {
-    required: true,
-    type: Boolean,
+  application: {
+    required: false,
+    type: Object,
+    default: null,
   },
   postApplicationUrl: {
     required: true,
@@ -115,6 +162,26 @@ const props = defineProps({
 });
 
 const isUserType = computed(() => props.posting.type === 'user');
+
+const statusLabel = computed(() => ({
+  open: "Under review",
+  accepted: "Accepted",
+  rejected: "Not selected",
+}[props.application?.status] ?? props.application?.status));
+
+const statusClass = computed(() => ({
+  open: "bg-yellow-50 text-yellow-700",
+  accepted: "bg-green-50 text-green-700",
+  rejected: "bg-gray-100 text-gray-600",
+}[props.application?.status] ?? "bg-gray-100 text-gray-600"));
+
+const progressPercent = computed(() => {
+  if (!props.application?.total_stages) {
+    return 0;
+  }
+
+  return Math.min(100, Math.round((props.application.current_position / props.application.total_stages) * 100));
+});
 
 const form = useForm({
   corporation_id: props.posting.corporation_id,
