+++ src/Http/Resources/JobPostingResource.php
@@ -4,10 +4,11 @@ namespace Seatplus\Web\Http\Resources;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\JsonResource;
+use Seatplus\Web\Models\Recruitment\Enlistment;
 use Seatplus\Web\Models\Recruitment\EnlistmentReviewRound;
 
 /**
- * @mixin \Seatplus\Web\Models\Recruitment\Enlistment
+ * @mixin Enlistment
  */
 class JobPostingResource extends JsonResource
 {
